third_party/blink/web_tests/platform/generic/fullscreen/full-screen-table-section-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.webkitFullscreenElement.id is 'test'

Tests that a table-section element can be used in full screen mode.

Test Element

Hit Enter to run if not in test runner.
